+++ src/hike/app_info.py
@@ -1,0 +1,36 @@
+"""Application metadata that should not depend on the Textual runtime."""
+
+##############################################################################
+# Local imports.
+from . import __version__
+
+##############################################################################
+APP_NAME = "Hike"
+APP_DESCRIPTION = "A Markdown browser for the terminal."
+APP_VERSION = __version__
+HELP_TITLE = f"{APP_NAME} v{APP_VERSION}"
+HELP_ABOUT = """
+`Hike` is a terminal-based Markdown viewer; it was created
+by and is maintained by [Dave Pearson](https://www.davep.org/); it is
+Free Software and can be [found on
+GitHub](https://github.com/davep/hike).
+"""
+HELP_LICENSE = """
+Hike - A Markdown viewer for the terminal.  \n    Copyright (C) 2025 Dave Pearson
+
+This program is free software: you can redistribute it and/or modify it
+under the terms of the GNU General Public License as published by the Free
+Software Foundation, either version 3 of the License, or (at your option)
+any later version.
+
+This program is distributed in the hope that it will be useful, but WITHOUT
+ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+more details.
+
+You should have received a copy of the GNU General Public License along with
+this program. If not, see <https://www.gnu.org/licenses/>.
+"""
+
+
+### app_info.py ends here